+++ CadEditor/Settings_Mermaid-Utils.cs
@@ -53,9 +53,9 @@ public static class MermaidUtils
       int layoutAddr = ConfigScript.getLayoutAddr(curActiveLayout);
       int width =  ConfigScript.getLevelWidth(curActiveLayout);
       int height = ConfigScript.getLevelHeight(curActiveLayout);
-      byte[] layer = new byte[width * height];
+      int[] layer = new int[width * height];
       for (int i = 0; i < width * height; i++)
-          layer[i] = (byte)((Globals.romdata[layoutAddr + i] + 1)%256);
+          layer[i] = (Globals.romdata[layoutAddr + i] + 1)%256;
       return new LevelLayerData(width, height, layer, null, null);
   }
 }
