+++ CadEditor/settings_flintstones_surprise_at_dinosaur_peak/Settings_Flintstones2-1-rom.cs
@@ -49,7 +49,7 @@ public class Data
   
   LevelLayerData getLayout(int levelNo)
   {
-    byte[] layer = new byte[1];
+    int[] layer = new int[1];
     layer[0] = 1;
     return new LevelLayerData(1, 1, layer);
   }
